test/requirements.txt
crcmod==1.7
pyserial==3.5
